refactor: move user to one component
